マイグレーションファイル変更 drop時のテーブル存在判定を行わなければrefresh時にエラーが出るため。

+++ database/migrations/2018_03_15_131446_mod_saved_designs_table_add_column_image.php
@@ -25,8 +25,10 @@ class ModSavedDesignsTableAddColumnImage extends Migration
      */
     public function down()
     {
-        Schema::table('saved_designs', function (Blueprint $table) {
-            $table->dropColumn('image');
-        });
+        if (Schema::hasTable('saved_designs')) {
+            Schema::table('saved_designs', function (Blueprint $table) {
+                $table->dropColumn('image');
+            });
+        }
     }
 }
